Add config

+++ liscripts/restart.sh
@@ -1,7 +1,45 @@
 #!/bin/bash
 
+if [ 1 == 2 ]
+then
 rm bench_out.txt
-./liscripts/reactive.sh ./plan1_ycsb5.json &
+./liscripts/reactive.sh ./plan1_ycsb.json &
+pid=$!
+while kill -0 "$pid" >/dev/null 2>&1; do
+	if tail -100 bench_out.txt | grep -q "with NaN ms avg latency" 
+	then
+		echo "Aborted!!!!! Restart"
+		pkill -f reactive		
+		pkill -f uniform
+		pkill -f plan
+		pkill -f sar
+		pkill -f ant
+		rm bench_out.txt
+		./liscripts/reactive.sh ./plan1_ycsb.json &
+		pid=$!
+	else
+		res=`tail -80 bench_out.txt | grep "ms avg latency" | awk -F 'with ' '{print $2}' | awk -F ' ' '{print $1}'`
+		ff=($res)
+		if [ ${#ff[@]} -gt 1 ] && [ ${ff[0]} == ${ff[-1]} ] && [ ${ff[0]} == ${ff[1]} ]
+		then
+			echo "Aborted!!!!! Restart"
+			pkill -f reactive		
+			pkill -f uniform
+			pkill -f plan
+			pkill -f sar
+			pkill -f ant
+			rm bench_out.txt
+			./liscripts/reactive.sh ./plan1_ycsb.json &
+			pid=$!
+		fi
+	fi
+	sleep 10
+done
+sleep 3
+fi
+
+rm bench_out.txt
+./liscripts/proactive.sh ./plan1_ycsb.json &
 pid=$!
 
 while kill -0 "$pid" >/dev/null 2>&1; do
@@ -9,13 +47,63 @@ while kill -0 "$pid" >/dev/null 2>&1; do
 	then
 		echo "Aborted!!!!! Restart"
 		pkill -f reactive		
+		pkill -f uniform
+		pkill -f plan
+		pkill -f sar
+		pkill -f ant
 		rm bench_out.txt
-		./liscripts/reactive.sh ./plan1_ycsb5.json &
+		./liscripts/proactive.sh ./plan1_ycsb.json &
 		pid=$!
+	else
+		res=`tail -80 bench_out.txt | grep "ms avg latency" | awk -F 'with ' '{print $2}' | awk -F ' ' '{print $1}'`
+		ff=($res)
+		if [ ${#ff[@]} -gt 1 ] && [ ${ff[0]} == ${ff[-1]} ] && [ ${ff[0]} == ${ff[1]} ]
+		then
+			echo "Aborted!!!!! Restart"
+			pkill -f reactive		
+			pkill -f uniform
+			pkill -f plan
+			pkill -f sar
+			pkill -f ant
+			rm bench_out.txt
+			./liscripts/proactive.sh ./plan1_ycsb.json &
+			pid=$!
+		fi
 	fi
 	sleep 10
 done
-sleep 3
+
+while kill -0 "$pid" >/dev/null 2>&1; do
+	if tail -100 bench_out.txt | grep -q "with NaN ms avg latency" 
+	then
+		echo "Aborted!!!!! Restart"
+		pkill -f reactive		
+		pkill -f uniform
+		pkill -f plan
+		pkill -f sar
+		pkill -f ant
+		rm bench_out.txt
+		./liscripts/proactive.sh ./plan1_ycsb.json &
+		pid=$!
+	else
+		res=`tail -80 bench_out.txt | grep "ms avg latency" | awk -F 'with ' '{print $2}' | awk -F ' ' '{print $1}'`
+		ff=($res)
+		if [ ${#ff[@]} -gt 1 ] && [ ${ff[0]} == ${ff[-1]} ] && [ ${ff[0]} == ${ff[1]} ]
+		then
+			echo "Aborted!!!!! Restart"
+			pkill -f reactive		
+			pkill -f uniform
+			pkill -f plan
+			pkill -f sar
+			pkill -f ant
+			rm bench_out.txt
+			./liscripts/proactive.sh ./plan1_ycsb.json &
+			pid=$!
+		fi
+	fi
+	sleep 10
+done
+
 exit
 
 #rm bench_out.txt
